Inject Offset into Code Generator Class
